+++ IntegrationBus/include/ib/mw/IComAdapter.hpp
@@ -11,6 +11,13 @@
 #include "ib/mw/fwd_decl.hpp"
 #include "ib/sim/fwd_decl.hpp"
 
+namespace spdlog {
+namespace details {
+    struct log_msg;
+}
+    class logger;
+}
+
 namespace ib {
 namespace mw {
 
@@ -67,6 +74,7 @@ public:
     virtual auto GetParticipantController() -> sync::IParticipantController* = 0;
     virtual auto GetSystemMonitor() -> sync::ISystemMonitor* = 0;
     virtual auto GetSystemController() -> sync::ISystemController* = 0;
+    virtual auto GetLogger() -> std::shared_ptr<spdlog::logger>& = 0;
 
     virtual void RegisterCanSimulator(sim::can::IIbToCanSimulator* busSim) = 0;
     virtual void RegisterEthSimulator(sim::eth::IIbToEthSimulator* busSim) = 0;
@@ -130,6 +138,9 @@ public:
     virtual void SendIbMessage(EndpointAddress from, const sync::ParticipantCommand& msg) = 0;
     virtual void SendIbMessage(EndpointAddress from, const sync::SystemCommand& msg) = 0;
 
+    virtual void SendIbMessage(EndpointAddress from, const logging::LogMsg& msg) = 0;
+    virtual void SendIbMessage(EndpointAddress from, logging::LogMsg&& msg) = 0;
+    
     /*! \brief Block until all bus messages are transmitted and acknowledged.
     *
     *   Some middleware implementations do not inherently ensure in-order transportation, so we