+++ IntegrationBus/include/ib/mw/logging/all.hpp
@@ -1,0 +1,1 @@
+#include "LoggingDatatypes.hpp"