+++ IntegrationBus/include/ib/mw/logging/fwd_decl.hpp
@@ -4,15 +4,11 @@
 
 namespace ib {
 namespace mw {
+namespace logging {
 
-struct EndpointAddress;
+struct SourceLoc;
+struct LogMsg;
 
-class IComAdapter;
-    
+} // namespace logging
 } // namespace mw
 } // namespace ib
-
-// also include subdirectories
-#include "sync/fwd_decl.hpp"
-#include "logging/fwd_decl.hpp"
-