+++ IntegrationBus/include/ib/mw/logging/spdlog.hpp
@@ -1,0 +1,5 @@
+#include "spdlog/spdlog.h"
+#include "spdlog/fmt/ostr.h"
+#ifdef SendMessage
+#undef SendMessage
+#endif